# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API for e-commerce ",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.7:8180/api"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "localhost"
        },
        {
            "url": "https://shopjozi.ir/api"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "List all Orders self user",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Store cart products to order",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "erp_code": {
                                            "description": "erp_code",
                                            "type": "string"
                                        },
                                        "quantity": {
                                            "description": "quantity",
                                            "type": "number"
                                        },
                                        "attr": {
                                            "description": "attr",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/save": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Store cart products to order",
                "description": "[suggest: {1, 0}, shipping_method: {TAXI, NORMAL}, payment_method: {HOME_DELIVERY, WALLET}]",
                "operationId": "8457da312b2f3c5f38390f58067e6c8e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "description": "suggest",
                                        "type": "string"
                                    },
                                    "suggest": {
                                        "description": "suggest",
                                        "type": "string"
                                    },
                                    "shipping_method": {
                                        "description": "TAXI|NORMAL",
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "description": "WALLET|HOME_DELIVERY",
                                        "type": "string"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "erp_code": {
                                                    "description": "erp_code",
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "description": "quantity",
                                                    "type": "number"
                                                },
                                                "attr": {
                                                    "description": "attr",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Display Order by orderId self user",
                "operationId": "3a2d78e76e9fc3f06bb879fb7d48b40d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/details/{order}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Display Order by details self user",
                "operationId": "be695870c65e838f34cfb50765c69476",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{order}/cancelled": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "order cancelled",
                "operationId": "076442f0f32a99ab8814be6b4015a8d6",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/orders/courier/cost": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Display Order cost",
                "operationId": "579e2bd0451f1212c489b80980ffd881",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List all Products",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "VISIT_COUNT",
                                "enum": [
                                    "VISIT_COUNT",
                                    "CHEAP",
                                    "EXPENSIVE",
                                    "DISCOUNT"
                                ]
                            }
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "subcategory",
                        "in": "query",
                        "description": "required send param subcategory with param category",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "bAAbNA5zDg0="
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}